Telegram Bot introduction
added a Telegram bot -> with this we can send alerts. Still open is how we calculate the alerts and what kind of alerts we send. In the future we could also implement helper functions -> An AI supporting and answerng quick questions, like how big the supply for tomorrow is or sth. like this